+++ modules_s/auth/challenge.c
@@ -186,7 +186,8 @@ static inline int challenge(struct sip_msg* _msg, str* _realm, int _qop,
  */
 int www_challenge(struct sip_msg* _msg, char* _realm, char* _qop)
 {
-	return challenge(_msg, (str*)_realm, (int)_qop, 401, MESSAGE_401, WWW_AUTH_CHALLENGE);
+	return challenge(_msg, (str*)_realm, (int)(long)_qop, 401, MESSAGE_401,
+						WWW_AUTH_CHALLENGE);
 }
 
 
@@ -195,7 +196,8 @@ int www_challenge(struct sip_msg* _msg, char* _realm, char* _qop)
  */
 int proxy_challenge(struct sip_msg* _msg, char* _realm, char* _qop)
 {
-	return challenge(_msg, (str*)_realm, (int)_qop, 407, MESSAGE_407, PROXY_AUTH_CHALLENGE);
+	return challenge(_msg, (str*)_realm, (int)(long)_qop, 407, MESSAGE_407, 
+						PROXY_AUTH_CHALLENGE);
 }
